+++ doc/src/changelog.md
@@ -14,6 +14,12 @@
   a garden's scope, the `environments` block in a tree's scope, the
   `environments` block in global configuration scope and, lastly, OS environment
   variables. The first entry found is used when expanding variable expressions.
+  ([#23](https://github.com/davvid/garden/pull/23))
+
+- Evaluation cycles (i.e. circular variable dependencies) are now prevented when
+  evaluating garden variables. The evaluation engine will now return empty strings
+  when a variable with a cyclical expression is evaluated.
+  ([#24](https://github.com/davvid/garden/pull/24))
 
 
 ## v1.2.1